Move it to other async test cases
